refactor: move reusable methods into `api/lib/`
