+++ api/items.add.js
@@ -2,6 +2,8 @@ import {
   getProjectCoreDataQuery,
   addIssueToProjectMutation,
 } from "./lib/queries.js";
+import { projectFieldsNodesToFieldsMap } from "./lib/project-fields-nodes-to-fields-map.js";
+import { itemFieldsNodesToFieldsMap } from "./lib/item-fields-nodes-to-fields-map.js";
 
 /**
  * @param {import("../").default} project
@@ -27,8 +29,14 @@ export default async function addItem(project, contentNodeId) {
     contentId: contentNodeId,
   });
 
-  const projectFields = toFields(project, projectNext.fields.nodes);
-  const fields = toItemFields(projectFields, item.fieldValues.nodes);
+  const projectFields = projectFieldsNodesToFieldsMap(
+    project,
+    projectNext.fields.nodes
+  );
+  const fields = itemFieldsNodesToFieldsMap(
+    projectFields,
+    item.fieldValues.nodes
+  );
 
   const common = {
     id: item.content.id,
@@ -61,96 +69,4 @@ export default async function addItem(project, contentNodeId) {
     isDraft: false,
     issueOrPullRequest,
   };
-}
-
-const READ_ONLY_FIELDS = [
-  "Assignees",
-  "Labels",
-  "Repository",
-  "Milestone",
-  "Linked Pull Requests",
-];
-
-/**
- * @param {import("..").default} project
- * @param {import("..").ProjectFieldNode[]} nodes
- * @returns {import("..").ProjectFieldMap}
- */
-function toFields(project, nodes) {
-  return nodes.reduce((acc, node) => {
-    if (READ_ONLY_FIELDS.includes(node.name)) {
-      return acc;
-    }
-
-    const key = fieldNameToInternalName(project, node.name);
-    acc[key] = {
-      id: node.id,
-      name: node.name,
-    };
-
-    // Settings is a JSON string. It contains view information such as column width.
-    // If the field is of type "Single select", then the `options` property will be set.
-    const settings = JSON.parse(node.settings);
-    if (settings?.options) {
-      acc[key].optionsById = settings.options.reduce((acc, option) => {
-        return {
-          ...acc,
-          [option.id]: option.name,
-        };
-      }, {});
-      acc[key].optionsByValue = settings.options.reduce((acc, option) => {
-        return {
-          ...acc,
-          [option.name]: option.id,
-        };
-      }, {});
-    }
-
-    return acc;
-  }, {});
-}
-
-/**
- * Returns internal name for a project field
- *
- * @param {import('..').default} project
- * @param {string} name
- * @returns string
- */
-function fieldNameToInternalName(project, name) {
-  const result = Object.entries(project.fields).find(
-    ([, value]) => value === name
-  );
-
-  if (!result) {
-    throw new Error(`Unknown column name: ${name}`);
-  }
-
-  return result[0];
-}
-
-/**
- * @param {import("..").ProjectFieldMap} projectFields
- * @param {import("..").ProjectFieldValueNode[]} nodes
- * @returns {Record<string, string>}
- */
-function toItemFields(projectFields, nodes) {
-  return Object.entries(projectFields).reduce(
-    (acc, [projectFieldName, projectField]) => {
-      const rawValue =
-        nodes.find((node) => node.projectField.id === projectField.id)?.value ||
-        null;
-
-      const value =
-        "optionsById" in projectField
-          ? projectField.optionsById[rawValue]
-          : rawValue;
-
-      return {
-        ...acc,
-        [projectFieldName]: value,
-      };
-    },
-    {}
-  );
 }
